Work done to code
After todays lab I realized my approach to the project was a bit wrong. I changed the code in all 3 files. The code is now working but only for 1 exercise. I am going to experiment with a for loop to allow for information to be stored for more than one exercise.